+++ firmware/gfx/weegfx.h
@@ -122,7 +122,7 @@ private:
 };
 
 inline void Graphics::setPixel(coord_t x, coord_t y) {
-  *(get_frame_ptr(x, y)) |= (0x1 << (y & 0x7));
+  *(get_frame_ptr(x, y)) |= static_cast<uint8_t>(0x1 << (y & 0x7));
 }
 
 inline void Graphics::drawAlignedByte(coord_t x, coord_t y, uint8_t byte) {
